+++ nomos-services/blend/src/edge/mod.rs
@@ -4,7 +4,9 @@ mod settings;
 use std::{fmt::Display, marker::PhantomData};
 
 use backends::BlendBackend;
-use nomos_blend_scheduling::message_blend::crypto::CryptographicProcessor;
+use nomos_blend_scheduling::{
+    message_blend::crypto::CryptographicProcessor, serialize_encapsulated_message,
+};
 use nomos_core::wire;
 use overwatch::{
     services::{
@@ -141,7 +143,5 @@ async fn handle_messages_to_blend<NodeId, Rng, Backend, RuntimeServiceId>(
     else {
         return;
     };
-    backend
-        .send(wire::serialize(&message).expect("Should not fail to serialize data payload."))
-        .await;
+    backend.send(serialize_encapsulated_message(&message)).await;
 }
