+++ nomos-services/blend/src/message.rs
@@ -1,3 +1,4 @@
+use nomos_blend_scheduling::EncapsulatedMessage;
 use serde::{Deserialize, Serialize};
 
 /// A message that is handled by [`BlendService`].
@@ -17,4 +18,24 @@ pub enum ServiceMessage<BroadcastSettings> {
 pub struct NetworkMessage<BroadcastSettings> {
     pub message: Vec<u8>,
     pub broadcast_settings: BroadcastSettings,
+}
+
+#[derive(Debug)]
+pub enum ProcessedMessage<BroadcastSettings> {
+    Network(NetworkMessage<BroadcastSettings>),
+    Encapsulated(Box<EncapsulatedMessage>),
+}
+
+impl<BroadcastSettings> From<NetworkMessage<BroadcastSettings>>
+    for ProcessedMessage<BroadcastSettings>
+{
+    fn from(value: NetworkMessage<BroadcastSettings>) -> Self {
+        Self::Network(value)
+    }
+}
+
+impl<BroadcastSettings> From<EncapsulatedMessage> for ProcessedMessage<BroadcastSettings> {
+    fn from(value: EncapsulatedMessage) -> Self {
+        Self::Encapsulated(Box::new(value))
+    }
 }
